Refactor FIDO2 code for reviewability
